fetch cvss score from cwe
